chore: rename Europeana APIs plugin
To europeana-apis.js, incorporating Nuxt config helpers
